api/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
pydantic==2.5.3
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pandas==2.0.3
requests==2.31.0
chromadb==0.5.5
sentence-transformers==3.0.1